Solution to Check Greater Than

+++ Check Greater Than@@67882a731b4392d8fc1577d6/code.c
@@ -2,8 +2,8 @@
 
 
 int main() {
-    float m,n;
-    scanf("%f",&m, &n);
+    double m,n;
+    scanf("%lf",&m, &n);
     if(m>n){
         printf("True");
     }
